Menambahkan fungsi daftar(), membuat Sistem Login, Logout dan Session

+++ hapus.php
@@ -1,4 +1,9 @@
 <?php
+    session_start();
+    if(!isset($_SESSION['login'])){
+        header('Location: login.php');
+        exit;
+    }
     require 'fungsi.php';
     $id = $_GET['id'];
     query("DELETE FROM mahasiswa WHERE id =$id");
